src/realtime_getdata/KKT_Module/requirements.txt
altgraph==0.17
anaconda==0.0.1.1
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
clr==1.0.3
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
et-xmlfile==1.0.1
future==0.18.2
h5py==3.1.0
idna==2.10
install==1.3.4
ipython==7.22.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
obs-websocket-py==0.5.3
openpyxl==3.0.5
pandas==1.2.1
parso==0.8.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
playsound==1.2.2
prompt-toolkit==3.0.18
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
requests==2.25.0
rsa==4.6
scikit-learn==0.24.1
scipy==1.6.0
shiboken2==5.15.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
websocket-client==1.2.1
wxPython==4.1.1
xlrd==1.2.0
